requirements.txt
streamlit
pandas
qrcode[pil]
pillow
openpyxl